+++ demo/spiders/newstracklive_重构.py
@@ -75,60 +75,42 @@ class newstracklive(scrapy.Spider):
         for test_page in test_page_list:
             page_list.append(test_page.find('a').get('href'))
         for i in page_list:
-            yield Request(i , callback=self.parse_1)
+            yield Request(i , callback=self.parse_2)
 
-    def parse_1(self , response ,  **kwargs):
+    def parse_2(self , response ,  **kwargs):
         page_soup = BeautifulSoup(response.text, 'lxml')
         category1 = page_soup.select('div.main-title-outer.pull-left div.main-title')[0].text.strip()
-        next_page_url = 'https://www.newstracklive.com/entertainment-news/television-news/' + \
-                        page_soup.select('div.nt_detailview div.ntdv_pagination > ul > li')[-1].find('a').get(
-                               'href')
-        this_page_url = response.url
-        j = 0
-        while j < 1 and next_page_url:
-            this_page_soup = BeautifulSoup(response.text, 'lxml')
-            news_url_list = this_page_soup.select(
-                'div.container div.col-sm-16.business.wow.fadeInDown.animated div.row div.col-md-16.col-sm-16 div.col-md-4.col-sm-8.col-xs-16 > div.topic.nt_topic > a')
-            last_new_url = news_url_list[-1].get('href')
-            last_new_time = time_font(BeautifulSoup(requests.get(last_new_url).text,'lxml').select('div.text-danger.sub-info-bordered div.time')[0].text)
-            if self.time == None or Util.format_time3(last_new_time) >= int(self.time):
-                for news_url in news_url_list:
-                    item = DemoItem()
-                    item = {
-                        'title': '','body': '','pub_time': '','category1': '','category2': '','images': [],'abstract': ''
-                    }
-                    item['category1'] = category1
-                    item['category2'] = category1
-                    yield Request(news_url, callback=self.parse_2 , meta={'item' : item})
-                j = j + 1
-                this_page_url = next_page_url
-                try:
-                    next_page_url = 'https://www.newstracklive.com/entertainment-news/television-news/' + \
-                                    this_page_soup.select('div.nt_detailview div.ntdv_pagination > ul > li')[-1].find(
-                                        'a').get('href')
-                except:
-                    print('没有下一页了')
-                    break
+        item = DemoItem()
+        item['category1'] = category1
+        item['category2'] = category1
+        for i in page_soup.select('div.col-md-4.col-sm-8.col-xs-16 div.topic.nt_topic a'):
+            yield Request(i.attrs['href'],callback=self.parse_3,meta={'item':item})
+        if page_soup.select('div.ntdv_pagination li')[-1].find('a').attrs['href']:
+            next_page = response.url + page_soup.select('div.ntdv_pagination li')[-1].find('a').attrs['href']
+            last_news_url = BeautifulSoup(requests.get(next_page).text,'lxml').select('div.col-md-4.col-sm-8.col-xs-16 div.topic.nt_topic a')[-1].attrs['href']
+            last_time = time_font(BeautifulSoup(requests.get(last_news_url).text,'lxml').select('div.time')[0].text)
+            if self.time == None or Util.format_time3(last_time) >= int(self.time):# 截止功能
+                #下一页
+                yield Request(next_page,callback=self.parse_2)
             else:
                 self.logger.info('时间截止')
-                break
 
-    def parse_2(self, response):
+    def parse_3(self, response):
         item = response.meta['item']
         new_soup = BeautifulSoup(response.text , 'lxml')
         item['title'] = \
         new_soup.select('div.sec-topic.nt_detailview.col-sm-16.wow.fadeInDown.animated div.col-sm-16.sec-info > h1')[
             0].text
+        item['body'] = ''
         for bodys in new_soup.find_all('p', style='text-align: justify;'):
             item['body'] += bodys.text
         item['abstract'] = new_soup.find_all('p', style='text-align: justify;')[0].text
-        try:
+        item['images'] = []
+        if len(new_soup.select(
+                'div.sec-topic.nt_detailview.col-sm-16.wow.fadeInDown.animated div.ntdv_imgcon > img')):
             new_images_list = new_soup.select(
                 'div.sec-topic.nt_detailview.col-sm-16.wow.fadeInDown.animated div.ntdv_imgcon > img')
             for new_images in new_images_list:
                 item['images'].append(new_images.get('src'))
-        except:
-            print("没有图片")
-
         item['pub_time'] = time_font(new_soup.select('div.text-danger.sub-info-bordered div.time')[0].text)
         yield item
